[Ref]: 검색 오버레이 컴포넌트를 SearchView로 분리
- HomeTabView 내 SearchOverlayView를 SearchView로 분리하여 파일 정리
- UI 구조는 동일하며 유지보수를 고려해 컴포넌트 분리
- 더 명확한 역할 분리를 통해 가독성과 재사용성 개선
